refactored floor
