+++ src/aframe/component/architecture-toolkit/floor.js
@@ -5,10 +5,8 @@
 import getSchema from './common/get-schema.js'
 import getMaterial from './common/get-material.js'
 import updateSchema from './common/update-schema.js'
-import generatePolygonBuffer from '../../../utils/data3d/buffer/get-polygon'
-import generateExtrusionBuffer from '../../../utils/data3d/buffer/get-extrusion'
-import generateNormals from '../../../utils/data3d/buffer/get-normals'
 import cloneDeep from 'lodash/cloneDeep'
+import floorData3d from '../../../scene/structure/parametric-objects/floor'
 
 export default {
 
@@ -18,7 +16,7 @@ export default {
 
   updateSchema: updateSchema,
 
-  update: function (oldData) {
+  update: async function (oldData) {
     var this_ = this
     var data = this_.data
 
@@ -26,19 +24,7 @@ export default {
     this.remove()
 
     // get defaults and
-    this.attributes = cloneDeep(data)
-
-    // get meshes and materials from el3d modules
-    var meshes = this.generateMeshes3d()
-
-    // clean up empty meshes to prevent errors
-    var meshKeys = Object.keys(meshes)
-    meshKeys.forEach(key => {
-      if (!meshes[key].positions || !meshes[key].positions.length) {
-        // console.warn('no vertices for mesh', key)
-        delete meshes[key]
-      }
-    })
+    let attributes = cloneDeep(data)
 
     // setup materials
     // defaults
@@ -58,16 +44,10 @@ export default {
       materials[mesh] = data[key]
     })
 
-    // fetch materials from mat library
-    Object.keys(materials).forEach(mat => {
-      materials[mat] = getMaterial(materials[mat])
-    })
+    attributes.materials=materials
 
     // construct data3d object
-    var data3d = {
-      meshes: meshes,
-      materials: materials
-    }
+    var data3d = await floorData3d(attributes)
 
     // create new one
     this_.mesh = new THREE.Object3D()
@@ -91,64 +71,5 @@ export default {
     }
   },
 
-  generateMeshes3d: function () {
-    var a = this.attributes
 
-    // 2d polygon vertices
-    var vertices = [ 0, 0, 0, a.w, a.l, a.w, a.l, 0 ]
-
-    // top polygon
-    var topPolygon = generatePolygonBuffer({
-      outline: vertices,
-      y: 0,
-      uvx: a.x,
-      uvz: a.z
-    })
-
-    // ceiling polygon
-    var ceilingPolygon
-    if (a.hasCeiling) {
-      ceilingPolygon = generatePolygonBuffer({
-        outline: vertices,
-        y: a.hCeiling,
-        uvx: a.x,
-        uvz: a.z,
-        flipSide: true
-      })
-    } else {
-      ceilingPolygon = {
-        vertices: new Float32Array(0),
-        uvs: new Float32Array(0)
-      }
-    }
-
-    // sides
-    var sides = generateExtrusionBuffer({
-      outline: vertices,
-      y: -a.h,
-      flipSide: true
-    })
-
-    // return meshes
-    return {
-      top: {
-        positions: topPolygon.vertices,
-        normals: generateNormals.flat(topPolygon.vertices),
-        uvs: topPolygon.uvs,
-        material: 'top'
-      },
-      sides: {
-        positions: sides.vertices,
-        normals: generateNormals.flat(sides.vertices),
-        uvs: sides.uvs,
-        material: 'side'
-      },
-      ceiling: {
-        positions: ceilingPolygon.vertices,
-        normals: generateNormals.flat(ceilingPolygon.vertices),
-        uvs: ceilingPolygon.uvs,
-        material: 'ceiling'
-      }
-    }
-  }
 }
